requirements.txt
django==4.2.11
psycopg==3.2.3
python-dotenv==1.0.1
django-split-settings
django-cors-headers==4.5.0
django-debug-toolbar==4.4.6
django-extensions==3.2.3
elasticsearch==8.16.0
elastic-transport==8.15.1
pydantic==2.10.2
pydantic_core==2.27.1

